requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.26.0
python-telegram-bot==20.8
APScheduler==3.10.4
SQLAlchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
fake-useragent==1.5.1
playwright==1.42.0
aiosqlite==0.20.0
